more refactoring, optimized structure, removed files
