+++ src/features/caption/config/nicheOptions.ts
@@ -1,0 +1,12 @@
+export const nicheOptions = [
+  { value: 'fashion', label: 'Fashion' },
+  { value: 'food', label: 'Food & Dining' },
+  { value: 'fitness', label: 'Fitness & Health' },
+  { value: 'tech', label: 'Technology' },
+  { value: 'travel', label: 'Travel' },
+  { value: 'beauty', label: 'Beauty & Wellness' },
+  { value: 'business', label: 'Business & Finance' },
+  { value: 'education', label: 'Education' },
+  { value: 'entertainment', label: 'Entertainment' },
+  { value: 'lifestyle', label: 'Lifestyle' },
+];
