+++ src/features/caption/hooks/useAuth.ts
@@ -1,5 +1,6 @@
 import { useSupabase } from '@/components/Providers';
 
+// logic for authentication
 export const useAuth = () => {
   const { session } = useSupabase();
   return { session };
